minor fixes to text progress bar
